Implement problem rejudging
This also allows the initial judging vs rejudging code to be deduplicated by adding an intermediate "not yet judged" state in the database. When a solution is initially submitted, an entry in the database is created with the "not yet judged" state, and then the same code as for rejudging is called.